Added flag to control tuple_cat implementation
It turns out that there's a small implementation detail that changes
the efficiency of tuple_cat, and this is whether tuple_cat uses a
forwarding tuple, or aggregates everything by value. With optimization
level O2 or above, aggregating everything by value should be more
efficient, but below that it may be more efficient to use a forwarding
tuple.

Signed-off-by: Alecto Irene Perez <perez.cs@pm.me>