Fix `-Wunused-parameter` warning
As I'm currently just including `tuplet.hpp` in my project and am not
using the full CMake capabilities, yet, all warnings that I've enabled
for my own project are enabled for tuplet as well.

This warning was just a bit annoying.  We can simply remove the argument
name to silence the warning.

Alternative would be:

```cpp
(void) tup;
```

or to wrap that in a custom macro such as `TUPLET_UNUSED(tup)`.